requirements.txt
accelerate==0.31.0
bitsandbytes==0.43.1
chainlit==1.1.202
datasets==2.19.2
errant==3.0.0
evaluate==0.4.2
faiss-cpu==1.8.0
gradio==4.36.1
happytransformer==3.0.0
huggingface_hub==0.23.3
ipython==8.12.3
language_tool_python==2.8
librosa==0.10.2.post1
nltk==3.8.1
moviepy==1.0.3
peft
pyannote.audio==3.2.0
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pydub==0.25.1
pytube==15.0.0
ragatouille==0.0.8.post2
Requests==2.32.3
scipy==1.13.1
sentence_splitter==1.4
sentence_transformers==2.7.0
spacy==3.7.5
soundfile==0.12.1
torch==2.3.1
transformers==4.44.2
fitz==0.0.1.dev2
frontend==0.0.3
pymupdf==1.24.5
yt_dlp
whisper-openai
openai==1.47.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
